fix window step subtraction
